+++ lib/easypost/models/brand.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+class EasyPost::Models::Brand < EasyPost::Models::EasyPostObject
+end