+++ lib/easypost/models/carbon_offset.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+class EasyPost::Models::CarbonOffset < EasyPost::Models::EasyPostObject
+end