+++ lib/easypost/models/report.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+class EasyPost::Models::Report < EasyPost::Models::EasyPostObject
+end