+++ lib/easypost/models/scan_form.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+class EasyPost::Models::ScanForm < EasyPost::Models::EasyPostObject
+end