+++ lib/easypost/models/tax_identifier.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+class EasyPost::Models::TaxIdentifier < EasyPost::Models::EasyPostObject
+end