+++ lib/easypost/models/tracker.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+class EasyPost::Models::Tracker < EasyPost::Models::EasyPostObject
+end