+++ lib/easypost/services/referral_customer.rb
@@ -99,6 +99,6 @@ class EasyPost::Services::ReferralCustomer < EasyPost::Services::Service
       },
     }
     referral_client = EasyPost::Client.new(api_key: referral_api_key)
-    referral_client.make_request(:get, 'credit_cards', EasyPost::Models::EasyPostObject, wrapped_params, 'beta')
+    referral_client.make_request(:post, 'credit_cards', EasyPost::Models::EasyPostObject, wrapped_params, 'beta')
   end
 end